Move local workers out of the gui-related modules
