+++ LDMP/localexecution/urbanchange.py
@@ -78,7 +78,7 @@ def compute_urban_change_summary_table(
         if clip_worker.success:
             log('Calculating summary table...')
             urban_summary_worker = worker.StartWorker(
-                calculate_urban.UrbanSummaryWorker,
+                UrbanSummaryWorker,
                 'calculating summary table (part {} of {})'.format(n + 1, len(wkts)),
                 str(output_indicator_tif),
                 urban_band_nums,
@@ -141,3 +141,91 @@ def save_summary_table(areas, populations, out_file):
             f"Error saving output table - check that {out_file!r} is accessible and "
             f"not already open. - {str(exc)}"
         )
+
+
+class UrbanSummaryWorker(worker.AbstractWorker):
+
+    def __init__(self, src_file, urban_band_nums, pop_band_nums, n_classes):
+        worker.AbstractWorker.__init__(self)
+
+        self.src_file = src_file
+        self.urban_band_nums = [int(x) for x in urban_band_nums]
+        self.pop_band_nums = [int(x) for x in pop_band_nums]
+        self.n_classes = n_classes
+
+    def work(self):
+        self.toggle_show_progress.emit(True)
+        self.toggle_show_cancel.emit(True)
+
+        src_ds = gdal.Open(self.src_file)
+
+        urban_bands = [src_ds.GetRasterBand(b) for b in self.urban_band_nums]
+        pop_bands = [src_ds.GetRasterBand(b) for b in self.pop_band_nums]
+
+        block_sizes = urban_bands[1].GetBlockSize()
+        xsize = urban_bands[1].XSize
+        ysize = urban_bands[1].YSize
+
+        x_block_size = block_sizes[0]
+        y_block_size = block_sizes[1]
+
+        src_gt = src_ds.GetGeoTransform()
+
+        # Width of cells in longitude
+        long_width = src_gt[1]
+        # Set initial lat ot the top left corner latitude
+        lat = src_gt[3]
+        # Width of cells in latitude
+        pixel_height = src_gt[5]
+
+        areas = np.zeros((self.n_classes, len(self.urban_band_nums)))
+        populations = np.zeros((self.n_classes, len(self.pop_band_nums)))
+
+        blocks = 0
+        for y in range(0, ysize, y_block_size):
+            if y + y_block_size < ysize:
+                rows = y_block_size
+            else:
+                rows = ysize - y
+            for x in range(0, xsize, x_block_size):
+                if self.killed:
+                    log("Processing of {} killed by user after processing {} out of {} blocks.".format(self.prod_out_file, y, ysize))
+                    break
+                self.progress.emit(100 * (float(y) + (float(x)/xsize)*y_block_size) / ysize)
+                if x + x_block_size < xsize:
+                    cols = x_block_size
+                else:
+                    cols = xsize - x
+
+                # Caculate cell area for each horizontal line
+                cell_areas = np.array([summary.calc_cell_area(lat + pixel_height*n, lat + pixel_height*(n + 1), long_width) for n in range(rows)])
+                # Convert areas from meters into hectares
+                cell_areas = cell_areas * 1e-4
+                cell_areas.shape = (cell_areas.size, 1)
+                # Make an array of the same size as the input arrays containing
+                # the area of each cell (which is identicalfor all cells ina
+                # given row - cell areas only vary among rows)
+                cell_areas_array = np.repeat(cell_areas, cols, axis=1)
+
+                # Loop over the bands (years)
+                for i in range(len(self.urban_band_nums)):
+                    urban_array = urban_bands[i].ReadAsArray(x, y, cols, rows)
+                    pop_array = pop_bands[i].ReadAsArray(x, y, cols, rows)
+                    pop_array[pop_array == -32768] = 0
+                    # Now loop over the classes
+                    for c in range(1, self.n_classes + 1):
+                        areas[c - 1, i] += np.sum((urban_array == c) * cell_areas_array)
+                        pop_masked = pop_array.copy() * (urban_array == c)
+                        # Convert population densities to persons per hectare
+                        # from persons per sq km
+                        pop_masked = pop_masked / 100
+                        populations[c - 1, i] += np.sum(pop_masked * cell_areas_array)
+
+                blocks += 1
+            lat += pixel_height * rows
+        self.progress.emit(100)
+
+        if self.killed:
+            return None
+        else:
+            return list((areas, populations))
